[issue-60] distinguish file and folder in resource tree
To be able to distinguish files and folder in the resource tree we add a type to the Resource class. When a new element is added to the opossum datastructure the corresponding path is added to the resources with the corresponding ResourceType. Files and snippets get the type FILE, packages FOLDER, the top-level Resource TOP_LEVEL and all other attributions get the type OTHER.

Signed-off-by: Meret Behrens <meret.behrens@tngtech.com>